% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_covid_data.R
\name{clean_covid_data}
\alias{clean_covid_data}
\title{Clean covid data}
\usage{
clean_covid_data(xl_data, outcome_nm, basecase, tx_levels)
}
\arguments{
\item{xl_data}{}

\item{outcome_nm}{}

\item{basecase}{}

\item{tx_levels}{}
}
\description{
Clean covid data
}
